Update contract_reflections.sol
This is an example of a solidity contract in 0.8.6 compiler. This works perfectly to create your own BSC tokens.

This contract has reflection tokenomics, in which a % of the buy and sell amount is reflected to all holders for passive income, like SafeMoon contract.

Contract is audited and approved by Certik.

For use with remix ethereum.